app/backend/requirements/dev.txt
-r base.txt

django_debug_toolbar
python-dotenv